add hooks/ folder and put useSign() and useSignAccessVP() there
