+++ src/context/useSymfoni.ts
@@ -111,54 +111,6 @@ export const useSymfoni = () => {
         [lazySigner, selectedSigner]
     );
 
-    const signAccessVP = async () => {
-        const url = process.env.REACT_APP_USE_LOCAL_ENVIROMENT === "true" ? "http://localhost:3004" : process.env.REACT_APP_BROK_HELPERS_URL;
-        return sign<{ jwt: string }[]>(
-            "symfoniID_accessVP",
-            "Gi Brønnøysundregistrene Forvalt applikasjonen tilgang til å gjøre spørringer på dine vegne",
-            [
-                {
-                    verifier: process.env.BROK_HELPERS_VERIFIER,
-                    access: {
-                        delegatedTo: {
-                            id: process.env.REACT_APP_PUBLIC_URL,
-                            name: "Brønnøysundregistrene Aksjeeierbok",
-                        },
-                        scopes: [
-                            {
-                                id: `${url}/captable/:captableAddress/shareholder/list`,
-                                name: "Lese alle aksjonærer",
-                            },
-                            {
-                                id: `${url}/captable/:captableAddress/shareholder/:shareholderId`,
-                                name: "Lese detaljer om aksjonærer",
-                            },
-                            {
-                                id: `${url}/unclaimed/list`,
-                                name: "Lese alle dine private aksjer",
-                            },
-                        ],
-                    },
-                },
-            ]
-        );
-    };
-
-    const sign = async <T>(method: string, message: string, params: any[]) => {
-        const request: SignatureRequest = {
-            message: message,
-            fn: async (_signer: Signer) => {
-                if (!("request" in _signer)) {
-                    debug(`connect(): !("request" in signer")`);
-                    return null;
-                }
-                return await _signer.request(method, params);
-            },
-        };
-        signatureRequestHandler.add([request]);
-        return signatureRequestHandler.resultsNT<T>();
-    };
-
     useAsyncEffect(
         async (isMounted) => {
             debug(`signerRequestAttempts: ${signerRequestAttempts}`);
@@ -202,7 +154,6 @@ export const useSymfoni = () => {
         address,
         chainId,
         closeSigner,
-        signAccessVP,
         ...contracts,
     };
 };
